+++ module/net/CHANGELOG.md
@@ -8,6 +8,8 @@ All notable changes to this project will be documented in this file.
 
 ### Added
 
+- Sign and verify hashes directly
+
 ### Changed
 
 ### Fixed